Re-do 3b 25
The original proof was valid but I tried using it as a reference when I was trying to do 3d 6, which is similar, and I didn't understand my solution, so I'm trying to make it clearer.
I didn't understand what was motivating me to do all these strange manipulations with Sv and Tv.
The point I was missing was that we can't define v in terms of a basis of V because V might not be finite-dimensional, but that doesn't mean we can't find Sv and Tv in terms of a basis of W.